+++ linen-modern/src/main/scala/x7c1/linen/modern/init/dev/CreateDummiesDelegatee.scala
@@ -4,7 +4,6 @@ import android.app.Activity
 import android.widget.Toast
 import x7c1.linen.glue.res.layout.DevCreateDummiesLayout
 import x7c1.linen.modern.accessor.LinenDatabase
-import x7c1.linen.modern.init.DummyFactory
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.patch.TaskAsync.async
 import x7c1.wheat.modern.decorator.Imports._
