+++ linen-modern/src/test/scala/x7c1/linen/modern/accessor/ChannelAccessorTest.scala
@@ -6,7 +6,7 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
-import x7c1.linen.modern.init.DummyFactory
+import x7c1.linen.modern.init.dev.DummyFactory
 
 
 @Config(manifest=Config.NONE)
