+++ linen-modern/src/test/scala/x7c1/linen/modern/accessor/SourceOpenHelperTest.scala
@@ -8,7 +8,7 @@ import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
 import x7c1.linen.modern.init.AccessorLoader.inspectSourceAccessor
-import x7c1.linen.modern.init.DummyFactory
+import x7c1.linen.modern.init.dev.DummyFactory
 
 @Config(manifest=Config.NONE)
 @RunWith(classOf[RobolectricTestRunner])
